feat: add unified glues binary and reorganize tui/server
