+++ bin/glues/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "glues"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+authors.workspace = true
+description = "Unified CLI entry point for the Glues toolkit"
+
+[dependencies]
+clap = { version = "4.5.4", features = ["derive"] }
+color-eyre = "0.6.3"
+glues-tui = { path = "../../tui", package = "glues-tui", default-features = false, features = ["cli"] }
+glues-server = { path = "../../server" }
+tokio = { version = "1.41.0", features = ["macros", "rt-multi-thread"] }
+
+[features]
+default = []
