+++ tui/src/main.rs
@@ -1,53 +1,10 @@
 #[cfg(not(target_arch = "wasm32"))]
-use glues::{App, config, logger, theme};
-
-#[cfg(not(target_arch = "wasm32"))]
-use {
-    clap::{Parser, ValueEnum},
-    color_eyre::Result,
-};
-
-#[cfg(not(target_arch = "wasm32"))]
-#[derive(Parser)]
-#[command(author, version, about)]
-struct Cli {
-    #[arg(long, value_enum, default_value_t = ThemeArg::Dark)]
-    theme: ThemeArg,
-}
-
-#[cfg(not(target_arch = "wasm32"))]
-#[derive(Clone, Copy, ValueEnum)]
-enum ThemeArg {
-    Dark,
-    Light,
-    Pastel,
-    Sunrise,
-    Midnight,
-    Forest,
-}
+use glues_tui::cli;
 
 #[cfg(not(target_arch = "wasm32"))]
 #[tokio::main]
-async fn main() -> Result<()> {
-    let cli = Cli::parse();
-
-    match cli.theme {
-        ThemeArg::Dark => theme::set_theme(theme::DARK_THEME),
-        ThemeArg::Light => theme::set_theme(theme::LIGHT_THEME),
-        ThemeArg::Pastel => theme::set_theme(theme::PASTEL_THEME),
-        ThemeArg::Sunrise => theme::set_theme(theme::SUNRISE_THEME),
-        ThemeArg::Midnight => theme::set_theme(theme::MIDNIGHT_THEME),
-        ThemeArg::Forest => theme::set_theme(theme::FOREST_THEME),
-    }
-
-    config::init().await;
-    logger::init().await;
-    color_eyre::install()?;
-
-    let terminal = ratatui::init();
-    let app_result = App::new().run(terminal).await;
-    ratatui::restore();
-    app_result
+async fn main() -> color_eyre::Result<()> {
+    cli::run_cli().await
 }
 
 #[cfg(target_arch = "wasm32")]
