+++ tui/web/index.html
@@ -15,7 +15,7 @@
       rel="rust"
       href="../Cargo.toml"
       data-cargo-manifest="../Cargo.toml"
-      data-target-name="glues"
+      data-target-name="glues_tui"
       data-target-kind="cdylib"
       data-cargo-args="--lib --no-default-features --features web"
     />
